requirements.txt
# Packages for Web Framework
fastapi==0.111.1
uvicorn==0.30.3
starlette==0.37.2
fastapi-cli==0.0.4
fastapi-utils==0.7.0

# Data Validation and Configuration
pydantic==2.8.2
pydantic-core==2.20.1
python-dotenv==1.0.1
pyyaml==6.0.1

# Networking and HTTP
requests==2.32.3
urllib3==2.2.2
httpx==0.27.0
httpcore==1.0.5
h11==0.14.0
httplib2==0.22.0
httptools==0.6.1
websockets==12.0
requests-oauthlib==2.0.0
oauthlib==3.2.2

# Asynchronous Programming
anyio==4.4.0
sniffio==1.3.1

# Database
SQLAlchemy==2.0.31

# Authentication and Authorization
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Google APIs
google-api-core==2.19.1
google-api-python-client==2.138.0
googleapis-common-protos==1.63.2

#  Other Dependencies
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
pip==23.2.1
proto-plus==1.24.0
protobuf==5.27.2
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-multipart==0.0.9
rich==13.7.1
rsa==4.9
shellingham==1.5.4
typing_extensions==4.12.2
typing-inspect==0.9.0
uritemplate==4.1.1
watchfiles==0.22.0